pyproject.toml
[tool.poetry]
name = "geolibs-dator"
packages = [
    { include = "dator" }
]
version = "0.0.3"
description = "GeoLibs Dator - A data extractor"
authors = ["Geographica <hello@geographica.com>"]
homepage = "https://github.com/GeographicaGS/GeoLibs-Dator"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
carto = "^1.4"
cartoframes = "^0.9.2"
google-cloud-bigquery = "^1.11"
marshmallow = "^2.19"
pandas = "^0.24.2"
psycopg2-binary = "^2.8.2"
PyYAML = "^5.1"
SQLAlchemy = "~1.2.19"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
jupyterlab = "^0.35.6"
pytest = "^4.5"
pytest-cov = "^2.7"
pytest-sugar = "^0.9.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
